\name{FisherButterflyCaptures}
\alias{FisherButterflyCaptures}
\docType{data}
\title{Fisher's butterfly data}
\description{The Fisher's butterfly data originally appeared in Fisher 1943.}
\references{
Fisher, R. A., Corbet, A. S., and Williams, C. B. ,1943, The Relation Between 
the Number of Species and the Number of Individuals in a Random Sample of an 
Animal Population, Journal of Animal Ecology, 12, 42-58.
}

\examples{
##load library
library(preseqR)

##load data that coming with the package.
data(FisherButterflyCaptures)
}
