\name{FisherButterflyHist}
\alias{FisherButterflyHist}
\docType{data}
\title{Fisher's butterfly data}
\description{Frequencies data of butterflies collected in the Malay peninsula
was from Fisher, R. A., Corbet, A. S., & Williams, C. B. (1943).}
\references{
Fisher, R. A., Corbet, A. S., and Williams, C. B. ,1943, The Relation Between 
the Number of Species and the Number of Individuals in a Random Sample of an 
Animal Population, Journal of Animal Ecology, 12, 42-58, Table 1,2.
}

\details{
    A two-column matrix.  
    The first column is the frequency \eqn{j = 1,2,\dots}; and the second column
    is \eqn{n_j}, the number of butterflies captured \eqn{j}
    times in the sample.
}  

\examples{
##load library
library(preseqR)

##load data
data(FisherButterflyHist)
}

\keyword{ data }
