% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer.R
\name{kmer.frac.curve}
\alias{kmer.frac.curve}
\title{Fraction of k-mers observed at least r times}
\usage{
kmer.frac.curve(n, k, read.len, seq, r = 2, mt = 20)
}
\arguments{
\item{n}{A two-column matrix. The first column is the frequency j
= 1,2,...; and the second column is N_j, the number of k-mers
observed exactly j times in the initial experiment. The first
column must be sorted in an ascending order.}

\item{k}{The number of nucleotides in a k-mer.}

\item{read.len}{The average length of a read.}

\item{seq}{The amount of nucleotides sequenced.}

\item{r}{A positive integer. Default is 1.}

\item{mt}{An positive integer constraining possible rational
function approximations. Default is 20.}
}
\value{
A two-column matrix. The first column is the amount of
    sequencing in an experiment. The second column is the estimate
    of the fraction of k-mers observed at least r times in the
    experiment.
}
\description{
'kmer.frac.curve' predicts the expected fraction of
    k-mers observed at least r times in a high-throughput
    sequencing experiment given the amount of sequencing
}
\details{
'kmer.frac.curve' is mainly designed for metagenomics to evaluate
how saturated is a given metagenomic sequencing data set.

'kmer.frac.curve' is the fast version of
'kmer.frac.curve.bootstrap'. The function does not provide the
confidence interval. To obtain the confidence interval along with
the estimates, one should use the function
'kmer.frac.curve.bootstrap'.
}
\examples{
# load library
library(preseqR)

# import data
data(SRR061157_k31)

# the fraction of 31-mers represented at least 10 times in an
# experiment when sequencing 1M, 10M, 100M, 1G, 10G, 100G, 1T
# nucleotides
kmer.frac.curve(n=SRR061157_k31, k=31, read.len=100,
                seq=10^(6:12), r=10, mt=20)
}
\references{
Efron, B., & Tibshirani, R. J. (1994). An introduction to the
bootstrap. CRC press.

Deng C, Daley T, Calabrese P, Ren J & Smith AD (2020).
Predicting the Number of Bases to Attain Sufficient Coverage in
High-Throughput Sequencing Experiments.
Journal of Computational Biology, 27(7), 1130-1143
}
\author{
Chao Deng \email{chaodeng@usc.edu}
}
