\name{preseqR.print2file}
\alias{preseqR.print2file}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to write results down to a file
}
\description{
The function writes down results generated by other functions in preseqR 
into a file, which could be easily read by users.
} 
\usage{
preseqR.print2file(X, filename = NULL, digit = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
	  A variable could be either a continued fraction with "CF" attribute or a
	   matrix with "matrix" attribute. See details for more information.  
  }
  \item{filename}{
	  A string is the name of output file.
  }
  \item{digit}{
	  A non-negative number controls the format of output numbers. It determines
	  the number of digits after the decimal point preserved.
  }
}

\details{
\code{preseqR.print2file} aims to help users process results when running 
functions in preseqR. Results of functions, including 
\code{preseqR.bootstrap.complexity.curve}, \code{preseqR.extrapolate.distinct}, 
\code{preseqR.interpolate.distinct}, \code{preseqR.ztnb.complexity.curve}, could
directly be an input for \code{preseqR.print2file}. The output of
\code{preseqR.continued.fraction.estimate} has two components. 
\code{preseqR.print2file} can write them down seperately.
}
\value{
	It returns either 0 or 1. 0 means normal. 1 means abnormal. If the returning
	value is 1, an error message will print on stderr.
}

\author{
	Chao Deng
}

\examples{
## load library
library(preseqR)

## import data
data(ShakespeareWordCounts)

## construct a continued fraction and estimate the number of observed word types
## if different quantities of Shakespeare's works are discovered.
result = preseqR.continued.fraction.estimate(ShakespeareWordCounts)
preseqR.print2file(result$continued.fraction, "test_continued_fraction.txt")
preseqR.print2file(result$yield.estimates, "test_yield_estimates.txt")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ print }
\keyword{ file }
