% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shakespeare.R
\docType{data}
\name{Shakespeare}
\alias{Shakespeare}
\title{Shakespeare's word count frequencies}
\format{
An object of class \code{data.frame}
\describe{
 \item{j:}{The count of appearances}
 \item{n_j:}{The number of words appearing this many times}
}
}
\usage{
data(Shakespeare)
}
\description{
The Shakespeare's word type frequencies data was from Efron, B., &
Thisted, R. (1976).
}
\details{
A two-column matrix.  The first column is the frequency j =
1,2,...; and the second column is n_j, the number of unique words
appeared j times in Shakespeare's work.
}
\examples{

# load library
library(preseqR)

# load this data
data(Shakespeare)

# check out the data
head(Shakespeare)
}
\references{
Efron, B., & Thisted, R. (1976). Estimating the number of unseen
species: How many words did Shakespeare know? Biometrika, 63(3),
435-447.
}
\keyword{Shakespeare}
