\name{ShakespeareWordCounts}
\alias{ShakespeareWordCounts}
\docType{data}
\title{Shakespeare's word type frequencies}
\description{The Shakespeare's word type frequencies data was from 
Efron, B., & Thisted, R. (1976).}
\references{
Efron, Bradley, and Ronald Thisted. "Estimating the number of unseen species: 
How many words did Shakespeare know?." Biometrika 63.3 (1976): 435-447, Table 1.
}

\examples{
##load library
library(preseqR)

##load data
data(ShakespeareWordCounts)
}
