\name{ShakespeareWordHist}
\alias{ShakespeareWordHist}
\docType{data}
\title{Shakespeare's word type frequencies}
\description{The Shakespeare's word type frequencies data was from 
Efron, B., & Thisted, R. (1976).}
\references{
Efron, B., & Thisted, R. (1976). Estimating the number of unseen species:
How many words did Shakespeare know?. Biometrika, 63(3), 435-447.
}

\examples{
##load library
library(preseqR)

##load data
data(ShakespeareWordHist)
}

\keyword{ data }
