% --- Source file: SRR1041830_5M_hist.txt --- 
\name{SRR1041830_5M_hist} 
\alias{SRR1041830_5M_hist} 
\title{Data set} 

\description{ 
The histogram has 20 rows and 2 columns. The first row is a header. Column 1 is
the frequency; column 2 is the number of distinct reads. The raw sequencing data
is from NCBI. All reads are first mapped to human genome and only five million
mapped reads are randomly sampled to build the histogram.
} 

\usage{data(SRR1041830_5M_hist)} 

\format{
This data contains the following variables:
    \itemize{
    \item frequencies: int
    \item distinct reads: int 
}
}
%\source{sequencing data see NCBI Accession: SRX386200}
\keyword{data}
