\name{FisherButterflyHist}
\alias{FisherButterflyHist}
\docType{data}
\title{Fisher's butterfly data}
\description{Frequencies data of butterflies collected in the Malay peninsula
was from Fisher, R. A., Corbet, A. S., & Williams, C. B. (1943).}
\references{
Fisher, R. A., Corbet, A. S., and Williams, C. B. ,1943, The Relation Between 
the Number of Species and the Number of Individuals in a Random Sample of an 
Animal Population, Journal of Animal Ecology, 12, 42-58, Table 1,2.
}

\examples{
##load library
library(preseqR)

##load data
data(FisherButterflyHist)
}

\keyword{ data }
