\name{preseqR.print2file}
\alias{preseqR.print2file}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to write results down to a file
}
\description{
preseqR.print2file write down variables generated by other functions in preseqR 
into a file, which could be read by users.
} 
\usage{
preseqR.print2file(X, prefix = '')
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
	  X could be either a continued fraction with "continued.fraction" attribute
	  or returning results from \code{preseqR.continued.fraction.estimate} or
	  \code{preseqR.bootstrap.complexity.curve}
  }
  \item{prefix}{
	  A string to indicate the prefix of created file name.
  }
}

\details{
preseqR.print2file aims to help users process results when running functions
in preseqR. It mainly takes returning results from either 
\code{preseqR.continued.fraction.estimate} or 
\code{preseqR.bootstrap.complexity.curve} as parameters and write them down 
into files. 
}
\value{
  \item{prefix_continued_fraction.txt}{
	  A file contains information about the constructed continued fraction. 
	  "prefix" in prefix_continued_fraction.txt will be replaced by its value
	  when the file is generated.
  }
  \item{yield.estimates}{
	  A file contains information about estimated number of distinct items given
	  various sizes of experiments.
  }
}

\author{
	Chao Deng
}

\examples{
## import data
data(SRR1041830_5M_hist)
## hist.count is a count vector of a histogram
hist.count = SRR1041830_5M_hist$distinct.reads
result = preseqR.continued.fraction.estimate(hist.count)
preseqR.print2file(result, "test")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ print }
\keyword{ file }
