\name{preseqR.print2file}
\alias{preseqR.print2file}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
A function to write results down to a file
}
\description{
The function writes down results generated by other functions in preseqR 
into a file, which could be easily read by users.
} 
\usage{
preseqR.print2file(X, prefix = '', digit = 0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
	  A variable could be either a continued fraction with "continued.fraction"
	  attribute or a matrix. See details for more information.  
  }
  \item{prefix}{
	  A string to indicate the prefix of created file name.
  }
  \item{digit}{
	  A non-negative number controls the format of output numbers. It determines
	  the number of digits after the decimal point preserved.
  }
}

\details{
preseqR.print2file aims to help users process results when running functions
in preseqR. It write down results from functions in preseqR into a file with
formate. Functions includes \code{preseqR.continued.fraction.estimate},
\code{preseqR.bootstrap.complexity.curve}, \code{preseqR.extrapolate.distinct}, 
\code{preseqR.interpolate.distinct}, \code{preseqR.ztnb.complexity.curve}.
}
\value{
  \item{prefix_continued_fraction.txt}{
	  A file contains information about the constructed continued fraction. 
	  "prefix" in prefix_continued_fraction.txt will be replaced by its value
	  when the file is generated.
  }
  \item{prefix_yield_estimates}{
	  A file contains information about estimated number of distinct items. 
	  "prefix" in prefix_continued_fraction.txt will be replaced by its value
	  when the file is generated.
  }
}

\author{
	Chao Deng
}

\examples{
## load library
library(preseqR)

## import data
data(ShakespeareWordCounts)

## convert a histogram into a count vector 
hist.count = vector(mode="numeric", length=max(ShakespeareWordCounts$frequencies))
hist.count[ShakespeareWordCounts$frequencies] = ShakespeareWordCounts$word.types

## construct a continued fraction and estimate the number of observed word types
## if different quantities of Shakespeare's works are discovered.
result = preseqR.continued.fraction.estimate(hist.count)
preseqR.print2file(result, "test")
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ print }
\keyword{ file }
