\name{ShakespeareWordCounts}
\alias{ShakespeareWordCounts}
\docType{data}
\title{Shakespeare's word type frequencies}
\description{The Shakespeare's word type frequencies data is from Efron, 
Bradley, and Ronald Thisted 1976 Table 1}
\references{
Efron, Bradley, and Ronald Thisted. "Estimating the number of unseen species: 
How many words did Shakespeare know?." Biometrika 63.3 (1976): 435-447.
}

\examples{
##load library
library(preseqR)

##load data that coming with the package.
data(FisherButterlyCaptures)
}
